backend/pyproject.toml
[project]
name = "praha-demo-backend"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi~=0.115.0",
    "uvicorn~=0.34.0",
    "httpx>=0.27.0",
    "pandas~=2.2.3",
    "numpy>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["services*", "routers*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
